configure: explicitly check for bison
The AC_PROG_YACC macro allows for 'bison -y', 'byacc' and
'yacc' but building with byacc fails with a syntax error:

  YACC     sql-parser.c
byacc: e - line 76 of "/home/berrange/src/virt/msitools/libmsi/sql-parser.y", syntax error
%define api.prefix {sql_}
^

Replace AC_PROG_YACC with an explicit check for bison -y
which is able to successfully generate the parser. An
added benefit of this custom check is that we can abort
configure with an error if it is not found instead of
using a hardcoded default of 'yacc' without checking if
it exists or not.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>