data: add validation of wxi files
The wxi files are prone to mistakes either at time of creation
or as packages change over time. Add a check script that can
validate for critical mistakes that will either break the
installer build, or prevent the binaries from being loaded
at runtime:

 - Missing DLLs refrenced by DLL/EXE files
 - Files in manifest which no longer exist on disk

This requires that the various mingw* RPMs are available
at time of "make check", otherwise the test is skipped.

Since RPM contents are continually changing, this test
is only expected to work against a Fedora rawhide
install.

Thus it is not wired into "make check" by default.
Instead it is run by "make -C data check-wxi"

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>