tools/wixl: Support use of FileKey in CustomAction
Working on project that builds a 32bit installer intended to be installed
on both 32bit and 64bit systems, to support this the CustomActions in the
installer need to be aware of the installed location.

MSI's support this through the use of `FileKey` which informs the MSI to
use the file installed on the system referenced by `FileKey`.

This pull request updates the fields of `WixCustomAction` to include the
field `FileKey` and updated the logic of `visit_custom_action` (in `builder.vala`)
to check for the validity of the field.

If the field has been set the custom action will be of the type `EXE_FILE`,
otherwise the existing behaviour of using `EXE_PROPERTY` remains.