wixl: Properly handle empty GUID in Component nodes
It's permissible for a ComponentId to be null (by setting the GUID
to an empty string). This results in Windows Installer ignoring the
component in all future operations (patch/uninstall) after
installing it once.

Without this patch, wixl creates broken installer packages when
specifying an empty GUID. This is because the string in the ComponentId
column comes out as "{}". As a result, when attempting to uninstall
the product, it is unregistered from the system but all its files
(even those with normal GUIDs) are still there.