Default Win64 to true on 64-bit arch
In WiX Toolset, the Component::Win64 attribute takes the correct value
depending on the `--arch` flag being built for. The official
documentation for Win64 says this:

> The default value is based on the platform set by the -arch switch to
> candle.exe or the InstallerPlatform property in a .wixproj MSBuild project:
> For x86 and ARM, the default value is 'no'. For x64, ARM64, and IA64,
> the default value is 'yes'.

This commit reproduces this behavior: If Win64 is unset, it defaults to
true when building for x64 and IA64 (WixL doesn't support ARM64 yet).