Strip scheduling condition nodes.
The Windows Installer parser for conditions does not like leading or
trailing spaces. This means that fairly common constructs like

    <Custom Action='dostuff' After='InstallServices'>
        (NOT Installed) AND (NOT REMOVE)
    </Custom>

will have the "wrong" condition `    (NOT Installed) AND (NOT REMOVE)`,
leading to the windows installer framework raising an error when
executing the action.

To avoid this, we simply strip the leading and trailing spaces of
conditions.