Remove default UI actions from AdminUISequence/InstallUISequence
ALl of these actions are defined in the respective dialog wxs files
such as WelcomeDlg, ExitDialog etc., so there is no need to
hardcode them.

Hardcoding causes issues with custom UIs that don't reference
all of the predefined dialogs, since there will be actions
in the sequence that don't exist.