Renamed libmsi to libmsi-1.0
As reported in bug report #48, the current library name conflicts with
libmsi.a which is the import library for the Microsoft Windows msi.dll
library API, that is present in various MinGW tool-chains.

This patch resolves the conflict by making a small modification to the
library name, bringing it into harmony with the name in the pkgconfig
file.

Signed-off-by: Joel Holdsworth <jholdsworth@nvidia.com>