wixl: Improving UI support
This patch makes it possible to use the slightly more advanced
WixUI_FeatureTree in addition to the current WixUI_Minimal flow.

The WixControl class has two added properties, Icon and Indirect,
mapping to the attribute flags ControlAttributes.ICON and
ControlAttributes.INDIRECT respectively. The primary property is the
Indirect one, which enables the BrowseDlg.wxs to correctly reference the
variable/path for choosing a custom install path. Icon was added for
completeness but I am unsure what effect it actually has.

The WixFeature class already had the AllowAdvertise and Absent
properties but they were not handled, meaning that no error or warning
was omitted, but the properties were never reflected in the MSI. I added
the two missing properties InstallDefault and TypicalDefault for
completeness. The four properties map to various parts of the features'
attributes, following the WiX documentation. Previously, the attributes
were set to 0 (aka None) regardless of properties in the xml.

After adding this attribute support, I was able to manipulate the
WixUI_FeatureTree files from UI.wixext to get a working installer.

To get the extended UI working, I had to replace all the localization
that is not supported by wixl, and move the Next/Previous buttons up in
the dialogs. And then it works.

Fixes: https://gitlab.gnome.org/GNOME/msitools/-/issues/62

Signed-off-by: Marc-André Lureau <marcandre.lureau@redhat.com>